# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Reserva de turnos odontológicos",
        "title": "Api de turnos Odontologicos",
        "termsOfService": "https://developers.ctd.com.ar/es_ar/terminos-y-condiciones",
        "contact": {
            "name": "API Support"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/pacientes": {
            "post": {
                "description": "Crea un nuevo paciente",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pacientes"
                ],
                "summary": "Crear un nuevo paciente",
                "parameters": [
                    {
                        "description": "Datos del paciente a crear",
                        "name": "paciente",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.Paciente"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Paciente creado correctamente",
                        "schema": {
                            "$ref": "#/definitions/domain.Paciente"
                        }
                    },
                    "400": {
                        "description": "Error de solicitud",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/pacientes/{idPaciente}": {
            "get": {
                "description": "Obtiene un paciente por su ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pacientes"
                ],
                "summary": "Obtener un paciente por ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID del paciente a obtener",
                        "name": "idPaciente",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paciente obtenido correctamente",
                        "schema": {
                            "$ref": "#/definitions/domain.Paciente"
                        }
                    },
                    "400": {
                        "description": "Error de solicitud",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "No se encontró el paciente",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Actualiza un paciente por su ID con todos los campos",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pacientes"
                ],
                "summary": "Actualizar un paciente por ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID del paciente a actualizar",
                        "name": "idPaciente",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Datos del paciente a actualizar",
                        "name": "paciente",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.Paciente"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paciente actualizado correctamente",
                        "schema": {
                            "$ref": "#/definitions/domain.Paciente"
                        }
                    },
                    "400": {
                        "description": "Error de solicitud",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "No se encontró el ID indicado",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Error al actualizar el paciente",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Elimina un paciente por su ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pacientes"
                ],
                "summary": "Eliminar un paciente por ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID del paciente a eliminar",
                        "name": "idPaciente",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "El paciente ha sido correctamente eliminado",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Error de solicitud",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "No se encontró el ID indicado",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "description": "Actualiza parcialmente un paciente por su ID con campos específicos",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pacientes"
                ],
                "summary": "Actualizar parcialmente un paciente por ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID del paciente a actualizar",
                        "name": "idPaciente",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Campos a actualizar",
                        "name": "paciente",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.Request"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paciente actualizado correctamente",
                        "schema": {
                            "$ref": "#/definitions/domain.Paciente"
                        }
                    },
                    "400": {
                        "description": "Error de solicitud",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "No se encontró el ID indicado",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Error al actualizar el paciente",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "domain.Paciente": {
            "type": "object",
            "required": [
                "apellidoPaciente",
                "dniPaciente",
                "domicilioPaciente",
                "fechaDeAltaPaciente",
                "nombrePaciente"
            ],
            "properties": {
                "apellidoPaciente": {
                    "type": "string"
                },
                "dniPaciente": {
                    "type": "string"
                },
                "domicilioPaciente": {
                    "type": "string"
                },
                "fechaDeAltaPaciente": {
                    "type": "string"
                },
                "idPaciente": {
                    "type": "integer"
                },
                "nombrePaciente": {
                    "type": "string"
                }
            }
        },
...
        "handler.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "handler.Request": {
            "type": "object",
            "properties": {
                "apellidoOdontologo": {
                    "type": "string"
                },
                "matriculaOdontologo": {
                    "type": "string"
                },
                "nombreOdontologo": {
                    "type": "string"
                }
            }
        }
    }
}